explicit cast after ++
git-svn-id: svn://svn.berlios.de/uoproxy/trunk@421 546381f1-97f9-0310-94fd-acf1737c2c57

+++ src/walk.c
@@ -130,7 +130,7 @@ connection_walk_request(struct linked_server *server,
 #endif
 
     walk = *p;
-    walk.seq = i->seq = state->seq_next++;
+    walk.seq = i->seq = (uint8_t)state->seq_next++;
     uo_client_send(server->connection->client.client, &walk, sizeof(walk));
 
     if (state->seq_next == 0)
